requirements.txt
google-generativeai
ollama